Address production problems

+++ server/db/migrate/20250506000945_add_missing_unique_indices.acts_as_taggable_on_engine.rb
@@ -7,22 +7,35 @@ class AddMissingUniqueIndices < ActiveRecord::Migration[7.1]
       execute('SET statement_timeout = 300000') # 5 minutes
     }
 
-    add_index ActsAsTaggableOn.tags_table, :name, unique: true, algorithm: :concurrently
+    # Add index to tags table name column if it doesn't exist already
+    unless index_exists?(ActsAsTaggableOn.tags_table, :name, unique: true, name: 'index_tags_on_name')
+      add_index ActsAsTaggableOn.tags_table, :name, unique: true, algorithm: :concurrently
+    end
 
     remove_index ActsAsTaggableOn.taggings_table, :tag_id if index_exists?(ActsAsTaggableOn.taggings_table, :tag_id)
-    remove_index ActsAsTaggableOn.taggings_table, name: 'taggings_taggable_context_idx'
-    add_index ActsAsTaggableOn.taggings_table,
-              %i[tag_id taggable_id taggable_type context tagger_id tagger_type],
-              unique: true, name: 'taggings_idx', algorithm: :concurrently
+    
+    if index_exists?(ActsAsTaggableOn.taggings_table, name: 'taggings_taggable_context_idx')
+      remove_index ActsAsTaggableOn.taggings_table, name: 'taggings_taggable_context_idx'
+    end
+    
+    # Only add the new index if it doesn't already exist
+    unless index_exists?(ActsAsTaggableOn.taggings_table, name: 'taggings_idx')
+      add_index ActsAsTaggableOn.taggings_table,
+                %i[tag_id taggable_id taggable_type context tagger_id tagger_type],
+                unique: true, name: 'taggings_idx', algorithm: :concurrently
+    end
   end
 
   def self.down
-    remove_index ActsAsTaggableOn.tags_table, :name
+    remove_index ActsAsTaggableOn.tags_table, :name if index_exists?(ActsAsTaggableOn.tags_table, :name)
 
-    remove_index ActsAsTaggableOn.taggings_table, name: 'taggings_idx'
+    remove_index ActsAsTaggableOn.taggings_table, name: 'taggings_idx' if index_exists?(ActsAsTaggableOn.taggings_table, name: 'taggings_idx')
 
     add_index ActsAsTaggableOn.taggings_table, :tag_id unless index_exists?(ActsAsTaggableOn.taggings_table, :tag_id)
-    add_index ActsAsTaggableOn.taggings_table, %i[taggable_id taggable_type context],
-              name: 'taggings_taggable_context_idx'
+    
+    unless index_exists?(ActsAsTaggableOn.taggings_table, name: 'taggings_taggable_context_idx')
+      add_index ActsAsTaggableOn.taggings_table, %i[taggable_id taggable_type context],
+                name: 'taggings_taggable_context_idx'
+    end
   end
 end
